In case of an error, don't delete the folder, allow us to debug

+++ 02 OCT Preprocess/runAllPreprocessing.m
@@ -138,8 +138,8 @@ catch ME
 		ME.stack(i) 
 	end 
 	disp(ME.message); 
-    
-    if (isUploadToCloud && deleteFolderAfterUpload) %Only manual cleaning
+   
+    if (false && isUploadToCloud && deleteFolderAfterUpload) %Only manual cleaning
         disp('Cleaning up local folder (as we alrady uploaded to the cloud)');
         %Delete local folder
         rmdir(SubjectFolderIn,'s');
